Update delete artifacts logic
We shouldn't try to delete artifacts if
GERRIT_CHANGE_NUMBER is not defined, otherwise we can have
incorrect path to artifacts that should be removed

+++ jenkins/jobs/builders/calico-delete-artifact-for-patchset.sh
@@ -2,5 +2,8 @@
 
 set -ex
 
+# do not perform any action if GERRIT_CHANGE_NUMBER is not defined
+[ -z "${GERRIT_CHANGE_NUMBER}" ] && exit 0
+
 # delete merged artifact
 curl -u "${ARTIFACTORY_LOGIN}:${ARTIFACTORY_PASSWORD}" -X DELETE "${ARTIFACTORY_URL}/${GERRIT_CHANGE_NUMBER}" || true
